Code cleanup #22

+++ radiotap.c
@@ -94,7 +94,7 @@ int ieee80211_radiotap_iterator_init(
 	int max_length, const struct ieee80211_radiotap_vendor_namespaces *vns)
 {
 	/* check the radiotap header can actually be present */
-	if (max_length < sizeof(struct ieee80211_radiotap_header))
+	if (max_length < (int)sizeof(struct ieee80211_radiotap_header))
 		return -EINVAL;
 
 	/* Linux only supports version 0 radiotap format */
